Fixed crash in cancel() if the job is not running.
svn path=/trunk/extragear/multimedia/k3b/; revision=608753

+++ src/rip/videodvd/k3bvideodvdrippingpreview.cpp
@@ -102,7 +102,7 @@ void K3bVideoDVDRippingPreview::generatePreview( const K3bVideoDVD::VideoDVD& dv
 
 void K3bVideoDVDRippingPreview::cancel()
 {
-  if( m_process->isRunning() ) {
+  if( m_process && m_process->isRunning() ) {
     m_canceled = true;
     m_process->kill();
   }
